+++ src/app/[service]/page.tsx
@@ -1,5 +1,5 @@
 import { fetchBookmarksByService } from '@/actions/bookmarks';
-import fetchServicesByPath from '@/actions/fetchServicesByPath';
+import { fetchServicesByPath } from '@/actions/services';
 import { isBookmarkEditable } from '@/actions/bookmarks';
 import { createClient } from '@/lib/supabase/server';
 import BookmarkGrid from './_components/BookmarkGrid';
