+++ src/types/registerService.ts
@@ -1,0 +1,6 @@
+export type RegisterService = {
+  serviceName: string;
+  serviceUrl: string;
+  userId?: string;
+  userEmail: string;
+};
